Managed to get db inside android/glass

+++ app/src/main/java/com/medialabamsterdam/checklistprototype/Polygon_contains_Point/Line.java
@@ -9,8 +9,8 @@ public class Line
 {
 	private final Point _start;
 	private final Point _end;
-	private float _a = Float.NaN;
-	private float _b = Float.NaN;
+	private double _a = Double.NaN;
+	private double _b = Double.NaN;
 	private boolean _vertical = false;
 
 	public Line(Point start, Point end)
@@ -39,10 +39,10 @@ public class Line
 	 */
 	public boolean isInside(Point point)
 	{
-		float maxX = _start.x > _end.x ? _start.x : _end.x;
-		float minX = _start.x < _end.x ? _start.x : _end.x;
-		float maxY = _start.y > _end.y ? _start.y : _end.y;
-		float minY = _start.y < _end.y ? _start.y : _end.y;
+		double maxX = _start.x > _end.x ? _start.x : _end.x;
+        double minX = _start.x < _end.x ? _start.x : _end.x;
+        double maxY = _start.y > _end.y ? _start.y : _end.y;
+        double minY = _start.y < _end.y ? _start.y : _end.y;
 
 		if ((point.x >= minX && point.x <= maxX) && (point.y >= minY && point.y <= maxY))
 		{
@@ -68,7 +68,7 @@ public class Line
 	 * 
 	 * @return The <b>A</b>
 	 */
-	public float getA()
+	public double getA()
 	{
 		return _a;
 	}
@@ -78,7 +78,7 @@ public class Line
 	 * 
 	 * @return The <b>B</b>
 	 */
-	public float getB()
+	public double getB()
 	{
 		return _b;
 	}
